Rename V. Rahul to Vemulapali Rahul
Restaurant Management System
My first step towards accomplishing the given task is this. As a customer walks in to a restaurant he will check for the availability of the table for the number of members present with him. So, my first step in accomplishing the given task is to check the availability of the tables in the restaurant for the customer and if the seats are available the customer is allotted the table and then bus boy serves him water later the customer is asked whether he would like to order the food and if he says yes he is given with a menu of which he has to choose and then the food is served to him later customer again gets an option whether he would like order some more food or not if he says yes the menu is again shown to him and if he says no he will be given a bill with applied taxes. This was my approach till now to complete the given task more changes are yet to be made. This is my work for the day more changes are to be done.