update to Restaurant Management System
In previous code we gave the menu details directly in the code by using println function now in this Update to the code we retrieve data i.e., Menu details from the file stored in the system. This reduces the risk of hard coding when the number of items in the menu increase. While executing this code we need to have a text file named "Menu" in the c drive of a system in which the code will be executed, this file contains the menu details. This is my work till date, still trying to update the code and complete the task within the deadline i.e., tomorrow. Thank you